feat: シリアル化
