+++ pubspec.yaml
@@ -10,6 +10,7 @@ dependencies:
   flutter:
     sdk: flutter
   flutter_svg: ^2.0.9
+  json_annotation: ^4.8.1
   yumemi_weather:
     git:
       url: https://github.com/yumemi-inc/flutter-training-template.git
@@ -17,9 +18,12 @@ dependencies:
       path: packages/yumemi_weather
 
 dev_dependencies:
+  build_runner: ^2.4.8
   flutter_test:
     sdk: flutter
+  json_serializable: ^6.7.1
   yumemi_lints: ^1.6.0
+
 
 flutter:
   uses-material-design: true
